refactor: break UpdateMethod decorator out into separate file
