+++ apps/api/src/domain/decorators/index.ts
@@ -2,3 +2,4 @@
  * TODO Extract this to a separate `web-of-knowledge` lib
  */
 export * from './aggregate-root.decorator';
+export * from './update-method.decorator';
